requirements.txt
anyio==3.7.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
channels==3.0.5
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.1
daphne==3.0.2
Django==4.2.2
django-cors-headers==4.1.0
django-environ==0.10.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
python-telegram-bot==20.3
pytz==2023.3
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
websocket-client==1.6.0
zope.interface==6.0
